+++ projects/igniteui-angular/src/lib/grids/hierarchical-grid/hierarchical-grid.virtualization.spec.ts
@@ -348,7 +348,7 @@ describe('IgxHierarchicalGrid Virtualization #hGrid', () => {
         fixture.componentInstance.data = fixture.componentInstance.generateData(10, 0);
         fixture.detectChanges();
 
-        fixture.componentInstance.rowIsland.onGridCreated.pipe(first(), delay(200)).subscribe(
+        fixture.componentInstance.rowIsland.gridCreated.pipe(first(), delay(200)).subscribe(
             async (args) => {
                 args.grid.data = fixture.componentInstance.generateData(10, 0);
                 await wait(200);
@@ -370,7 +370,7 @@ describe('IgxHierarchicalGrid Virtualization #hGrid', () => {
         expect((hierarchicalGrid.verticalScrollContainer.getScroll().children[0] as HTMLElement).offsetHeight).toEqual(561);
     });
 
-    it('should emit onScroll and onDataPreLoad on row island when child grid is scrolled.', async () => {
+    it('should emit onScroll and dataPreLoad on row island when child grid is scrolled.', async () => {
         const ri = fixture.componentInstance.rowIsland;
         const firstRow = hierarchicalGrid.dataRowList.toArray()[0];
         (firstRow.nativeElement.children[0] as HTMLElement).click();
@@ -382,7 +382,7 @@ describe('IgxHierarchicalGrid Virtualization #hGrid', () => {
 
 
         spyOn(ri.onScroll, 'emit').and.callThrough();
-        spyOn(ri.onDataPreLoad, 'emit').and.callThrough();
+        spyOn(ri.dataPreLoad, 'emit').and.callThrough();
 
 
         // scroll down
@@ -393,7 +393,7 @@ describe('IgxHierarchicalGrid Virtualization #hGrid', () => {
         fixture.detectChanges();
 
         expect(ri.onScroll.emit).toHaveBeenCalled();
-        expect(ri.onDataPreLoad.emit).toHaveBeenCalled();
+        expect(ri.dataPreLoad.emit).toHaveBeenCalled();
     });
 });